Improve handling of turn off mode.
Change name to 'Everything Off' rather than 'Turn Everything Off'. The
verb is inconsistent with the rest of the mode names and looks odd as
a label.

Move the off mode to the top of select menu where it will be easier to
find.

Remove the turn off mode from the list of modes in the turned off menu.
Looks odd there.