Add a link to cycle the receiver power for Fire TV
For some reason, my Fire TV sometimes has video but no sound. It
usually happens when I switch to the Fire TV input after the Fire
TV wakes up. Cycling the power on the receiver seems to fix it.
So, add a link that turns the receiver off and then on again.