Check status less often
I've been having issues with the server locking up. I think it
is happening somewhere in the communications with the receiver
over socket. Rather than attempt to fix the issue, I am simply
increasing the time before refreshing the status from the
receiver and on each client (by a lot). That should be a cheap
way to reduce the frequency of the lockup.