Make volume slider set relative volume
When using socket controls for the receiver, you can query the
exact volume and set the exact volume. However, the socket is
not super reliable. (I've had it lock up and not be responsive.)
When using IR, you cannot query the volume, and changing the
volume is not exact.

To avoid problems with the slider being stuck at the wrong end,
make the slider set a relative volume. The slider is always set
to the middle at rest, and sliding to left or right always sets
up or down.