Update codes to new equipment
Gave up on any socket-based control. Everything is now through
IR.

Changed around the control_central to make it easier to configure
new devices. Now there is a data structure that is easily
updated to switch out modes. This could be in a separate config
file, but for me it's just as easy to edit the source.